Fix logic for generating bytecode to reflect upstream changes in bcc.
https://github.com/iovisor/bcc/commit/8300c7bdd7f8599bdeac830cfed05ec8fd51616d
appears to have changed the way fds are vended out as part of
`bpf_pseudo_fd()`. In the new scheme, fds for BPF maps are added to a
`std::map` using negative numbers as placeholders (-1, -2, -3...), so
when they are enumerated for the purpose of being assigned *actual* fds
(once again, starting from 3 because 0/1/2 are stdin/stdout/stderr),
they are enumerated as ...-3, -2, -1.

This threw off the logic we used when extracting the uses of `ld_pseduo`
and parameterizing them when generating `generated_bytecode.h` and
`generated_bytecode.rs`. (Specifically, for generated functions that
took multiple BPF maps, the order was reversed.)

This commit fixes things by using `BPF.get_table()` after the bytecode
has been generated to get the final fd for the table by reading its
`.map_fd` field.