Update Rust code to reflect upstream changes in libbpf.h.
Note this is analogous to the fixes we did for opensnoop.c in:
https://github.com/bolinfest/opensnoop-native/commit/2993767a739f1f6339b9873f347a1b6a82dfa5b6

Though one thing I failed to mention on that diff is why the `-lelf`
argument was added (to both the C and Rust builds).
Without it, the build failed with something like:

```
  = note: /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_nextscn'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_rawdata'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `gelf_getshdr'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `gelf_getrel'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_getdata'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `gelf_getsym'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_getscn'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_begin'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_memory'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_end'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_strptr'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `gelf_getehdr'
          /usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/libbpf.so: undefined reference to `elf_version
```

Googling for `libbpf.so: undefined reference to `elf_nextscn'` turned up:
https://github.com/netoptimizer/prototype-kernel/issues/7

From there, I followed a link to:
https://github.com/netoptimizer/prototype-kernel/commit/db6bf73d6ed07689407bb2504cc58c7e1a7f8849

And on that page, I expanded the content of the `Makefile` until I saw:

```
LDFLAGS= -lelf
```

So I just tried adding `-lelf` and everything worked!

I don't know if it has anything to do with using a
`/usr/lib/x86_64-linux-gnu/libbpf.so` (or really, a
`/usr/lib/x86_64-linux-gnu/libbpf.so.0.9.0`) that I built from source
and installed with `make install` rather than an "official" version via
a package manager.